Fix filter dropdown layout issues in IE.
Part of MCKIN-1775.

Sometimes on first load of the discussion page, there would be
some extra padding on the filter dropdown items in IE.

The only way I was able to make that weird issue go away was setting
font-size to zero on the <li> elements.

Fortunately, that doesn't seem to have any negative side-effects,
because the content of each filter is wrapped in an <a> element,
and one of the existing CSS rules sets an explicit font-size
on <a> elements.