Temporary hack to test ooyala xblock optimization.
Units that use the ooyala xblock load very slowly.

Apros already loads the ooyala player JS library, so there is
no need to load another version for the ooyala xblock.

This is a temporary hack to test how much of a speed-up using
the already loaded ooyala player library brings.
If the performance improvement is significant,
we will implement a more proper solution.