MCKIN-4053: Notify group members on file submission.
Once user uploads any group work deliverable, he/she and all group
members received an email to confirm receipt.
A new email receipt is sent each time a new file is uploaded (i.e.,
replaced).

This is done by listening to the upload_complete event emitted by the
GWv2 xblock, and POST-ing a message with the submission_id to a new
handler that sends a confirmation email to all members of the group.